2.1 OBservableとはなにか
